Carry the double frame delta into sprite uploads and drop a dead JSON include
The sprite and billboard upload paths in both PALs still took the frame
delta as float after the clock moved to double; MSVC refuses the narrowing
under /WX, so every sprite scene failed the shipping build while clang-cl
stayed silent. The eight signatures take double, matching the elapsed
clocks they feed.

The sprite and node-particle units included ts_runtime.hpp without using
it, which pulled nlohmann's JSON header into scenes that reach no JSON;
the static shipping tree only carries that include directory behind the
loader:gltf, loader:babylon and data:json gate, so doom could not compile.
The templates no longer emit the include.

+++ native/src/pal_dawn_billboard.hpp
@@ -473,7 +473,7 @@ inline void upload_dawn_billboard_pass(
     DawnBillboardPass& pass,
     const std::array<float, 16>& view_projection,
     const std::array<float, 16>& view,
-    float delta_ms) {
+    double delta_ms) {
     const BillboardSystemRecord& system =
         engine.billboard_systems[pass.system.value];
